shop_api/requirements.txt
django
djangorestframework
django-filter
python-decouple
psycopg2-binary
drf-yasg
Pillow
python-slugify
djangorestframework-simplejwt
